Sitne ispravke sonarcloud bagova

+++ Klinicki-centar/src/main/java/MRSISA/Klinicki/centar/domain/Lekar.java
@@ -114,7 +114,7 @@ public class Lekar {
 	}
 	
 	public double izracunajProsecnuOcenu() {
-		int suma = 0;
+		double suma = 0;
 		int i = this.ocene.size();
 		if(i == 0) {
 			return 3.0;
